Add new feature to update player's money
A function 'UpdateMoney' was added to JoueurService and the Joueur Controller has been updated to include this. A default money amount of 1000 is set when creating a new player. A save changes command was added in the end of PartieService 'EndPartie' method. The 'Resultat' property in the 'Partie' entity is also updated to be non-nullable.